Tract OCoin balance changes
Hopefully this will help to track down remaining bugs in the points system.
It is possible that the points history info can be obtained from some logs,
but this works as well.

Thinking about this more maybe this could have used the new messaging system
instead of creating a brand new scheme, but whatever...